shuffling fixtures
