+++ src/encoded/tests/test_views.py
@@ -152,10 +152,10 @@ def test_load_sample_data(
 
 
 def test_abstract_collection(testapp, experiment):
-    #TODO: ASK_BEN how to get experiment to function as catch all
+    # TODO: ASK_BEN how to get experiment to function as catch all
     pass
-    #testapp.get('/experiment/{accession}'.format(**experiment))
-    #testapp.get('/expermient/{accession}'.format(**experiment))
+    # testapp.get('/experiment/{accession}'.format(**experiment))
+    # testapp.get('/expermient/{accession}'.format(**experiment))
 
 
 @pytest.mark.slow
@@ -164,7 +164,7 @@ def test_load_workbook(workbook, testapp, item_type, length):
     # testdata must come before testapp in the funcargs list for their
     # savepoints to be correctly ordered.
     res = testapp.get('/%s/?limit=all' % item_type).maybe_follow(status=200)
-    #TODO ASK_BEN about inherited collections i.e. protocol
+    # TODO ASK_BEN about inherited collections i.e. protocol
     assert len(res.json['@graph']) == length
